@@ -30,14 +30,14 @@
         ]
     },
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.2",
         "psr/http-message": "^1",
-        "webignition/web-resource-interfaces":">=0.11,<1",
+        "webignition/web-resource-interfaces":"^1",
         "webignition/internet-media-type-interface":"^0.1",
         "webignition/internet-media-type":"^1.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5",
+        "phpunit/phpunit": "^7",
         "squizlabs/php_codesniffer": "^3",
         "mockery/mockery": "^1",
         "webignition/web-resource-testing-tools":"^0.1"
